[Feature] fix #20: adjust homepage to single page
