--- app/controllers/contact_controller.rb
@@ -1,4 +1,0 @@
-class ContactController < ApplicationController
-  def index
-  end
-end
