--- app/helpers/contact_helper.rb
@@ -1,2 +1,0 @@
-module ContactHelper
-end
