--- app/helpers/work_helper.rb
@@ -1,48 +1,0 @@
-module WorkHelper
-  def tech_icon(icon)
-    case icon
-    when "Ruby on Rails"
-      "devicon-rails-plain"
-    when "HTML5"
-      "devicon-html5-plain"
-    when "CSS3"
-      "devicon-css3-plain"
-    when "JavaScript"
-      "devicon-javascript-plain"
-    when "React"
-      "devicon-react-plain"
-    when "Python"
-      "devicon-python-plain"
-    when "Tailwindcss"
-      "devicon-tailwindcss-plain"
-    when "Bootstrap"
-      "devicon-bootstrap-plain"
-    when "MySQL"
-      "devicon-mysql-plain"
-    when "MongoDB"
-      "devicon-mongodb-plain"
-    when "Firebase"
-      "devicon-firebase-plain"
-    when "Git"
-      "devicon-git-plain"
-    when "Postman"
-      "devicon-postman-plain"
-    when "WordPress"
-      "devicon-wordpress-plain"
-    when "Docker"
-      "devicon-docker-plain"
-    when "Ruby"
-      "devicon-ruby-plain"
-    when "PostgreSQL"
-      "devicon-postgresql-plain"
-    when "Redux"
-      "devicon-redux-plain"
-    when "Express"
-      "devicon-express-original"
-    when "NodeJS"
-      "devicon-nodejs-plain"
-    when "NextJS"
-      "devicon-nextjs-plain"
-    end
-  end
-end
