+++ app/views/components/_footer.html.erb
@@ -5,9 +5,6 @@
 
   <div class="links">
     <%= link_to "Home", root_path, class: "link" %>
-    <%= link_to "About", about_path, class: "link" %>
-    <%= link_to "Work", root_path, class: "link" %>
-    <%= link_to "Contact", root_path, class: "link" %>
     <%= link_to "Blogs", root_path, class: "link" %>
   </div>
