+++ app/views/components/_navbar.html.erb
@@ -5,9 +5,6 @@
 
   <div class="navbar">
     <%= link_to "Home", root_path, class: "link #{current_page?(root_path) ? 'link__active' : ''}" %>
-    <%= link_to "About", about_path, class: "link #{current_page?(about_path) ? 'link__active' : ''}" %>
-    <%= link_to "Work", work_path, class: "link #{current_page?(work_path) ? 'link__active' : ''}" %>
-    <%= link_to "Contact", contact_path, class: "link #{current_page?(contact_path) ? 'link__active' : ''}" %>
     <%= link_to "Blogs", root_path, class: "link" %>
   </div>
 </nav>
