--- app/views/contact/index.html.erb
@@ -1,13 +1,0 @@
-<div class="main-container">
-  <section class="hero-section">
-    <div class="heading">
-      <span class="bracket">[</span>
-      <h1 class="name text-center">CONTACT</h1>
-      <span class="bracket">]</span>
-    </div>
-  </section>
-
-  <div class="line-break"></div>
-
-  <%= render partial: "components/contact_me_form" %>
-</div>
