--- spec/requests/about_spec.rb
@@ -1,11 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "Abouts", type: :request do
-  describe "GET /index" do
-    it "returns http success" do
-      get "/about/index"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-end
